Create data_visualizer.py
To use this program, simply run the script and follow the prompts to enter the filename of the CSV file containing the health scores and ages of the chicken, as well as the filename to save the chart image. The program will then generate a line chart of the chicken's health over time and save it to the specified file.