Create Schedule.py
example of a program written in Python that can help automate the process of tracking a chicken's health over time using the chicken.py and life.py programs. This program uses a schedule library to schedule the running of the chicken.py program at regular intervals, and the data_visualizer.py program to generate a line chart of the chicken's health over time.

----

To use this program, simply run the script and it will automatically run the chicken.py program at the specified interval and save the output to a CSV file. It will also generate a line chart of the chicken's health over time once per day using the data_visualizer.py program. The program uses the schedule library to schedule the running of these tasks, so it will continue running indefinitely until stopped manually.

Note: This is just one example of how this automation program could be implemented, and there are many other ways it could be customized and improved depending on the specific needs/use case.